🐛 remove retired skills surface from settings
