--- libs/frontend/features/settings/src/lib/sections/skills-section/skills-section.component.html
@@ -1,47 +1,0 @@
-<wo-section-heading
-	title="Skills"
-	subtitle="Browse and manage skills across org, department, project, and personal scopes. Promote a personal skill to share it with your team."
-/>
-
-<table class="wo-table wo-skills__table">
-	<thead>
-		<tr>
-			<th>Name</th>
-			<th>Scope</th>
-			<th>Version</th>
-			<th>Digest</th>
-			<th>Status</th>
-			<th></th>
-		</tr>
-	</thead>
-	<tbody>
-		@for (skill of skills(); track skill.name) {
-			<tr>
-				<td class="wo-mono wo-skills__name">{{ skill.name }}</td>
-				<td><wo-scope-chip [label]="skill.scope" [color]="scopeColor(skill.scope)" /></td>
-				<td class="wo-mono wo-skills__muted">{{ skill.version }}</td>
-				<td class="wo-mono wo-skills__muted">{{ skill.digest }}</td>
-				<td>
-					@if (skill.status === "pending-promotion") {
-						<wo-scope-chip label="pending" color="#A0855A" />
-					} @else {
-						<span class="wo-mono wo-skills__active">active</span>
-					}
-				</td>
-				<td class="wo-skills__actions">
-					@if (skill.scope === personalScope && skill.status !== "pending-promotion") {
-						<button type="button" class="wo-skills__promote">Promote →</button>
-					}
-					@if (skill.status === "pending-promotion") {
-						<button type="button" class="wo-skills__cancel">Cancel</button>
-					}
-				</td>
-			</tr>
-		}
-	</tbody>
-</table>
-
-<button type="button" class="wo-dashed-button">
-	<i class="pi pi-plus"></i>
-	Install skill from registry
-</button>
