--- libs/frontend/features/settings/src/lib/sections/skills-section/skills-section.component.scss
@@ -1,43 +1,0 @@
-.wo-skills {
-	&__table {
-		margin-bottom: 24px;
-	}
-
-	&__name {
-		font-size: 13px;
-	}
-
-	&__muted {
-		font-size: 11px;
-		color: var(--muted-foreground);
-	}
-
-	&__active {
-		font-size: 10px;
-		color: var(--wo-ok);
-		opacity: 0.8;
-	}
-
-	&__actions {
-		text-align: right;
-	}
-
-	&__promote {
-		font-size: 11px;
-		color: var(--accent);
-		white-space: nowrap;
-
-		&:hover {
-			opacity: 0.7;
-		}
-	}
-
-	&__cancel {
-		font-size: 11px;
-		color: var(--muted-foreground);
-
-		&:hover {
-			opacity: 0.7;
-		}
-	}
-}
