+++ libs/frontend/features/settings/src/lib/settings-page/settings-page.component.html
@@ -29,9 +29,6 @@
 				@case (sections.Awareness) {
 					<wo-awareness-section />
 				}
-				@case (sections.Skills) {
-					<wo-skills-section />
-				}
 				@case (sections.Channels) {
 					<wo-channels-section />
 				}
