+++ libs/frontend/state/settings/adapter/src/lib/settings-mapper.util.ts
@@ -1,27 +1,7 @@
-import { DatasetAccess, EgressDomain, ScopeLevel, SkillRow } from "@opencrane/core";
+import { DatasetAccess, EgressDomain, ScopeLevel } from "@opencrane/core";
 
 import { AccountProfile, AccountProfileUpdate, AwarenessContractInfo, BudgetSpend, PodIdentity } from "./settings-gateway.types";
-
-/**
- * Wire shape of the OpenCrane `Tenant` identity fields the Account section reads.
- *
- * A local projection of the generated `Tenant` contract — only the fields the
- * mapping consumes are declared, all optional as the contract marks them.
- */
-export interface AccountTenantWire
-{
-	/** Stable tenant identifier. */
-	name?: string;
-
-	/** Human-readable display name. */
-	displayName?: string;
-
-	/** Org-managed email address. */
-	email?: string;
-
-	/** Team the tenant belongs to. */
-	team?: string;
-}
+import type { AccountTenantPatch, AccountTenantWire, BudgetSpendWire, DatasetsWire, EffectiveContractWire, PodTenantWire, PolicyWire } from "./settings-mapper.types";
 
 /**
  * Map a wire `Tenant` onto the Account read model.
@@ -41,23 +21,6 @@ export function _MapAccountProfile(wire: AccountTenantWire, fallbackName: string
 		email: wire.email ?? "",
 		department: wire.team ?? ""
 	};
-}
-
-/**
- * Patch shape for `PUT /tenants/{name}` — the writable identity fields only.
- *
- * A local projection of the contract's update body; the section never writes
- * `email` (org-managed) or `name` (the immutable path key), so neither appears.
- * Keys are present only when the corresponding edit was supplied, so the request
- * stays a minimal partial update.
- */
-export interface AccountTenantPatch
-{
-	/** New display name, when changing it. */
-	displayName?: string;
-
-	/** New team, when changing it. */
-	team?: string;
 }
 
 /**
@@ -103,36 +66,6 @@ function _ScopeFromWire(scope: string | undefined): ScopeLevel
 }
 
 /**
- * Wire shape of `GET /tenants/{name}` for the Pod & Session section.
- *
- * A local projection of the generated `Tenant` contract; all fields optional as
- * the contract marks them.
- */
-export interface PodTenantWire
-{
-	/** Stable tenant identifier. */
-	name?: string;
-
-	/** Human-readable display name. */
-	displayName?: string;
-
-	/** Org-managed email address. */
-	email?: string;
-
-	/** Team the tenant belongs to. */
-	team?: string;
-
-	/** Lifecycle phase. */
-	phase?: string;
-
-	/** Ingress host the pod is reachable on. */
-	ingressHost?: string;
-
-	/** ISO creation timestamp. */
-	createdAt?: string;
-}
-
-/**
  * Map a wire `Tenant` onto the Pod identity read model.
  *
  * Pure and DI-free. Optional wire fields collapse to empty strings;
@@ -155,23 +88,6 @@ export function _MapPodIdentity(wire: PodTenantWire, fallbackName: string): PodI
 }
 
 /**
- * Wire shape of `GET /ai-budget/{tenantName}/spend`.
- *
- * Local projection; fields optional as the contract marks them.
- */
-export interface BudgetSpendWire
-{
-	/** Monthly spend ceiling in USD. */
-	monthlyLimitUsd?: number;
-
-	/** Spend so far this month in USD. */
-	currentSpendUsd?: number;
-
-	/** Alert band (`ok` | `warning` | `exceeded`). */
-	budgetAlertState?: string;
-}
-
-/**
  * Map a wire spend payload onto the Budget read model.
  *
  * Pure and DI-free. Missing figures collapse to `0`; an unrecognised alert band
@@ -190,21 +106,6 @@ export function _MapBudgetSpend(wire: BudgetSpendWire): BudgetSpend
 }
 
 /**
- * Wire shape of the typed fields on `GET /tenants/{name}/effective-contract`.
- *
- * Only the flat identity fields are projected; the nested `awareness`/`mcp`/
- * `skills` blocks are opaque in the pinned contract.
- */
-export interface EffectiveContractWire
-{
-	/** Stable contract identifier. */
-	contractId?: string;
-
-	/** Resolved contract version string. */
-	contractVersion?: string;
-}
-
-/**
  * Map a wire effective-contract payload onto the Awareness identity read model.
  *
  * Pure and DI-free. Missing fields collapse to empty strings.
@@ -217,24 +118,6 @@ export function _MapAwarenessContract(wire: EffectiveContractWire): AwarenessCon
 		contractId: wire.contractId ?? "",
 		contractVersion: wire.contractVersion ?? ""
 	};
-}
-
-/**
- * Wire shape of `GET /tenants/{name}/datasets` — dataset names grouped by scope.
- */
-export interface DatasetsWire
-{
-	/** Org-scope dataset names. */
-	org?: string[];
-
-	/** Team (dept)-scope dataset names. */
-	team?: string[];
-
-	/** Project-scope dataset names. */
-	project?: string[];
-
-	/** Personal-scope dataset names. */
-	personal?: string[];
 }
 
 /**
@@ -264,69 +147,6 @@ export function _MapDatasetAccess(wire: DatasetsWire): DatasetAccess[]
 		}
 	}
 	return rows;
-}
-
-/**
- * Wire shape of a `GET /skills/catalog` row.
- *
- * Local projection of the fields the Skills table renders.
- */
-export interface SkillCatalogWire
-{
-	/** Skill name. */
-	name?: string;
-
-	/** Version string. */
-	version?: string;
-
-	/** OCI digest. */
-	digest?: string;
-
-	/** Scope token (`org` | `team` | `project` | `personal`). */
-	scope?: string;
-
-	/** Publication status (`draft` | `published` | `deprecated`). */
-	status?: string;
-}
-
-/**
- * Map skill-catalogue rows onto the Skills table read model.
- *
- * Pure and DI-free. The contract's publication status maps onto the table's
- * status vocabulary (`published` → `active`, `draft` → `pending-promotion`),
- * with any other value passed through. Missing strings collapse to `—`.
- *
- * @param wire - Catalogue rows as returned by the API.
- */
-export function _MapSkills(wire: SkillCatalogWire[]): SkillRow[]
-{
-	return wire.map(function mapRow(row: SkillCatalogWire): SkillRow
-	{
-		const status = row.status === "published"
-			? "active"
-			: row.status === "draft" ? "pending-promotion" : row.status ?? "active";
-		return {
-			name: row.name ?? "—",
-			scope: _ScopeFromWire(row.scope),
-			version: row.version ?? "—",
-			digest: row.digest ?? "—",
-			status
-		};
-	});
-}
-
-/**
- * Wire shape of a `GET /policies` row.
- *
- * Local projection of the fields the egress allowlist renders.
- */
-export interface PolicyWire
-{
-	/** Policy name. */
-	name?: string;
-
-	/** Allowed egress domains. */
-	domains?: string[];
 }
 
 /**
